requirements-vercel.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
google-generativeai>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.26.0
scipy>=1.11.0
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1
# Note: FFmpeg is required for pydub audio processing
# Install separately: conda install -c conda-forge ffmpeg
